server: move the ts import code out of server.go
Release note: None
